added categories slider
